requirements.txt
blosc2==2.0.0
grpcio==1.66.0
gym==0.26.2
numpy==1.23.4
protobuf==5.27.3
psutil==5.9.0